fix/tests: contract

+++ contracts/course/course_registry/src/functions/get_courses_by_instructor.rs
@@ -26,7 +26,7 @@ pub fn get_courses_by_instructor(env: &Env, instructor: Address) -> Vec<Course> 
         }
 
         id += 1;
-        if id > crate::schema::MAX_LOOP_GUARD {
+        if id > crate::schema::MAX_LOOP_GUARD as u128 {
             break; // safety limit
         }
     }
